refactor: integrate JWT auth with existing User model and models.py
