--- Backend/app/models/users.py
@@ -1,17 +1,0 @@
-from sqlalchemy import Column, String, DateTime
-from sqlalchemy.dialects.postgresql import UUID
-import uuid
-from datetime import datetime
-from ..db.db import Base
-
-class User(Base):
-    __tablename__ = "users"
-
-    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    username = Column(String, unique=True, index=True, nullable=False)
-    email = Column(String, unique=True, index=True, nullable=False)
-    hashed_password = Column(String, nullable=False)
-    role = Column(String, default="creator") # 'creator' or 'brand'
-    profile_image = Column(String, nullable=True)
-    bio = Column(String, nullable=True)
-    created_at = Column(DateTime, default=datetime.utcnow)
